chore: split evaluation/ICR, worksheets, analytics, bulk-diagnostic routes out of index.ts
Pure mechanical extraction, no behavior change. Completes the split
started in batch-a/b/c - this is the final batch, closing all 15
Phase 4 issues.

- routes/evaluation.ts - the biggest single group: ICR blue-pen
  filter/evaluate-pdf/cloud-config/evaluate-cloud (Python subprocess +
  cv2/OCR pipeline), plus POST /api/evaluation/submit, GET
  /api/evaluation/reports, GET /api/evaluation/:studentId/history.
  Extracted via exact-line sed capture rather than manual retyping,
  given the size (~1000 lines) - preserves every byte unchanged.
- routes/analytics.ts - GET /api/analytics, GET
  /api/analytics/superadmin (incl. the buildMongoFilter helper used
  only by the superadmin route).
- routes/diagnosticBulk.ts - POST /api/diagnostic/bulk,
  GET /api/diagnostic/bulk/:jobId/progress|download,
  GET /api/diagnostic/student/:studentId/answer-key,
  POST /api/diagnostic/single.
- routes/worksheets.ts recovered here too: the initial sed-based
  extraction of the evaluation/ICR block accidentally consumed the
  worksheets routes sitting in between in the original file (a single
  sed delete spanned a wider range than the two capture files). Caught
  via a `grep` sanity pass before committing - confirmed every route
  path from the pre-split file maps to exactly one location post-split
  - and recovered the already-correct extraction from the batch-c
  branch rather than re-deriving it.
- This branch was proactively rebased onto batch-c (rather than main)
  before pushing, since it also touches routes/worksheets.ts - avoids
  the same overlapping-edit conflict batch-b hit against batch-a.

backend/src/index.ts is now 126 lines, down from the original 3566.

Verified: tsc --noEmit clean at every stage (initial split, post-fix,
post-rebase). Live-tested three times against fresh scratch MongoDB
seeds - most thoroughly after the rebase: 14 endpoints across all 15
route groups (auth, tickets, logbook, geo, classes, admin, teachers,
schools, interventions, best-practices, students, analytics x2,
evaluation) all 200; re-confirmed #209's State Admin scoping fix
intact; ran the full worksheets/generate personalized-question
pipeline end-to-end twice (different cycles); confirmed icr/filter,
icr/cloud-config, diagnostic/bulk progress, and diagnostic/student
answer-key all return real application-level responses (validation
errors / "not found" JSON), not generic route-missing 404s - proving
each handler is genuinely reached, not silently dropped.

Closes #126, closes #127, closes #129
