+++ backend/src/index.ts
@@ -33,6 +33,9 @@ import { registerInterventionRoutes } from './routes/interventions';
 import { registerBestPracticeRoutes } from './routes/bestPractices';
 import { registerStudentRoutes } from './routes/students';
 import { registerWorksheetRoutes } from './routes/worksheets';
+import { registerEvaluationRoutes } from './routes/evaluation';
+import { registerAnalyticsRoutes } from './routes/analytics';
+import { registerDiagnosticBulkRoutes } from './routes/diagnosticBulk';
 import { randomUUID } from 'crypto';
 import fs from 'fs';
 import bcrypt from 'bcrypt';
@@ -79,1740 +82,10 @@ registerStatsRoutes(app);
 
   registerStudentRoutes(app);
 
-  // ICR Blue-Pen Filter Stage (standalone — runs only the cv2 blue-pen
-  // isolation, no OCR). Returns the filtered image as a data URL so the
-  // frontend can preview the black-on-white filtered result before the
-  // ~3-second OCR step. This makes the blue-pen filter visible to the
-  // user instead of happening invisibly inside a single round-trip.
-  app.post('/api/icr/filter', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { imageDataUrl } = req.body || {};
-        if (!imageDataUrl || !imageDataUrl.startsWith('data:')) {
-          return res.status(400).json({ error: 'imageDataUrl is required and must be a data URL.' });
-        }
-
-        // Accept raster images (PNG/JPEG/WebP) AND PDFs. The blue-ink filter
-            // operates on pixels, so PDFs are rasterized to PNG page 1 first.
-            // The image regex has 2 capture groups (mime subtype + b64); the PDF
-            // regex has 1 (just b64) — keep that asymmetry in mind below.
-            const imgMatch = /^data:image\/([a-zA-Z0-9+.-]+);base64,(.+)$/.exec(imageDataUrl);
-            const pdfMatch = /^data:application\/pdf;base64,(.+)$/.exec(imageDataUrl);
-            if (!imgMatch && !pdfMatch) {
-              return res.status(400).json({ error: 'imageDataUrl must be base64-encoded PNG/JPEG/WebP image or PDF.' });
-            }
-            let ext: string;
-                let b64: string;
-                const isPdf = !!pdfMatch;
-                if (isPdf) {
-                  ext = 'pdf';
-                  b64 = pdfMatch![1];
-                } else {
-                  ext = imgMatch![1] === 'jpeg' ? 'jpg' : imgMatch![1];
-                  b64 = imgMatch![2];
-                }
-                const buf = Buffer.from(b64, 'base64');
-        if (buf.length === 0) {
-          return res.status(400).json({ error: 'imageDataUrl decoded to zero bytes.' });
-        }
-        // Cap at 8 MB to match the evaluate-pdf endpoint's existing limit.
-        if (buf.length > 8 * 1024 * 1024) {
-          return res.status(413).json({ error: 'File too large (max 8 MB).' });
-        }
-
-        const tempDir = path.join(AI_SERVICES_DIR, 'scratch');
-        fs.mkdirSync(tempDir, { recursive: true });
-        const stamp = Date.now();
-        const inputPath = path.join(tempDir, `filter_${stamp}_in.${ext}`);
-        // After this point, the path the blue-ink filter reads from is `filterInputPath`.
-        // For images, it's the raw uploaded file; for PDFs, it's the rasterized PNG.
-        let filterInputPath = inputPath;
-            let pdfRasterizedPath: string | null = null;
-            const outputPath = path.join(tempDir, `filter_${stamp}_out.jpg`);
-
-        try {
-          fs.writeFileSync(inputPath, buf);
-
-          // PDF path: rasterize page 1 to PNG, then point filterInputPath at the PNG.
-          if (isPdf) {
-            const rasterScript = path.join(AI_SERVICES_DIR, 'scripts', 'pdf_rasterize.py');
-            const { execFileSync: execPdf } = await import('child_process');
-            const pdfOut = path.join(tempDir, `filter_${stamp}_raster.png`);
-            let pdfStdout: string;
-            try {
-              pdfStdout = execPdf(
-                PYTHON_BIN,
-                [rasterScript, inputPath, pdfOut],
-                { cwd: AI_SERVICES_DIR, timeout: 30000, encoding: 'utf8' }
-              );
-            } catch (e: any) {
-              return res.status(500).json({
-                success: false,
-                error: `PDF rasterization failed: ${e?.message || e}`,
-              });
-            }
-            const pdfLine = pdfStdout.trim().split('\n').filter(Boolean).pop() || '{}';
-            let pdfResult: any = {};
-            try {
-              pdfResult = JSON.parse(pdfLine);
-            } catch {
-              return res.status(500).json({
-                success: false,
-                error: `PDF rasterizer returned non-JSON: ${pdfStdout.slice(0, 300)}`,
-              });
-            }
-            if (!pdfResult.success) {
-              return res.status(500).json({ success: false, error: pdfResult.error || 'PDF rasterization failed.' });
-            }
-            filterInputPath = pdfOut;
-                        pdfRasterizedPath = pdfOut;
-                      }
-
-          const { execFileSync } = await import('child_process');
-          const scriptPath = path.join(AI_SERVICES_DIR, 'scripts', 'bluepen_filter.py');
-          const stdout = execFileSync(
-            PYTHON_BIN,
-            [scriptPath, filterInputPath, outputPath],
-            { cwd: AI_SERVICES_DIR, timeout: 30000, encoding: 'utf8' }
-          );
-          // Last non-empty line is the JSON result.
-          const jsonLine = stdout.trim().split('\n').filter(Boolean).pop() || '{}';
-          let parsed: any = {};
-          try {
-            parsed = JSON.parse(jsonLine);
-          } catch {
-            return res.status(500).json({ success: false, error: `Filter returned non-JSON: ${stdout.slice(0, 300)}` });
-          }
-          if (!parsed.success) {
-            return res.status(500).json({ success: false, error: parsed.error || 'Filter failed.' });
-          }
-          const filteredBuf = fs.readFileSync(outputPath);
-          const filteredDataUrl = `data:image/jpeg;base64,${filteredBuf.toString('base64')}`;
-          return res.json({
-            success: true,
-            imageDataUrl: filteredDataUrl,
-            bluePixelRatio: parsed.blue_pixel_ratio,
-            bluePixelCount: parsed.blue_pixel_count,
-            imageSize: parsed.image_size,
-            // Tell the client the input was a PDF so it can show a one-time
-            // "rasterized from PDF" note if it wants. Pure informational.
-            sourceType: isPdf ? 'pdf' : 'image',
-            // Pass the temp output path so the OCR step can read the same file
-            // without re-running the filter. (Frontend currently ignores this
-            // and re-uploads the data URL — both work; this is just an
-            // optimization for server-side chaining later.)
-            filteredPath: outputPath,
-          });
-        } catch (err: any) {
-          const msg = err?.message || String(err);
-          console.error('[icr-filter] failed:', msg);
-          return res.status(500).json({ success: false, error: msg });
-        } finally {
-          // Clean up the input; leave outputPath around briefly so the OCR
-          // endpoint could pick it up if it wanted (filteredPath). For now
-          // the frontend re-uploads the data URL, so outputPath is also safe
-          // to delete.
-          try { fs.unlinkSync(inputPath); } catch { /* noop */ }
-          if (pdfRasterizedPath) { try { fs.unlinkSync(pdfRasterizedPath); } catch { /* noop */ } }
-          try { fs.unlinkSync(outputPath); } catch { /* noop */ }
-        }
-      });
-
-  // ICR Answer Sheet Scanner (Single or Bulk Class Evaluation for PDF & Image uploads with EasyOCR)
-  app.post('/api/icr/evaluate-pdf', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { classId, studentId, pdfBase64, fileBase64, filename } = req.body;
-    const inputBase64 = fileBase64 || pdfBase64;
-    if (!inputBase64) {
-      return res.status(400).json({ error: 'fileBase64 or pdfBase64 is required.' });
-    }
-
-    // Fast path: no classId → single-image OCR. Just run EasyOCR once and
-    // return a flat answer list keyed by position (q_1, q_2, ...). No
-    // student/class lookup, no bulk evaluation. Used by the new two-stage
-    // ICR flow (frontend's IcrTwoStageScan component) which doesn't have
-    // a class context at scan time.
-    if (!classId) {
-      const tempDir = path.join(AI_SERVICES_DIR, 'scratch');
-      fs.mkdirSync(tempDir, { recursive: true });
-      const ext = path.extname(filename || 'worksheet.jpg') || '.jpg';
-      const tempFilePath = path.join(tempDir, `scan_noclass_${Date.now()}_file${ext}`);
-      const cleanBase64 = inputBase64.includes(',') ? inputBase64.split(',')[1] : inputBase64;
-      fs.writeFileSync(tempFilePath, Buffer.from(cleanBase64, 'base64'));
-      try {
-        const { execFileSync } = await import('child_process');
-        const scriptPath = path.join(AI_SERVICES_DIR, 'scripts', 'ocr.py');
-        const output = execFileSync(PYTHON_BIN, [scriptPath, tempFilePath, 'SCAN', '1'], {
-          cwd: AI_SERVICES_DIR,
-          env: { ...process.env, PYTHONIOENCODING: 'utf-8' },
-          timeout: 60000,
-          maxBuffer: 10 * 1024 * 1024,
-        });
-        const ocrResult = JSON.parse(output.toString());
-        const tokens = ocrResult?.evaluation?.extractedTokens || ocrResult?.extracted_tokens || [];
-        const rawText = ocrResult?.evaluation?.rawOcrText || ocrResult?.raw_text || '';
-        const detectedNumbers = ocrResult?.evaluation?.detectedNumbers || ocrResult?.digits_found || [];
-        // Build a flat answers map: q_1, q_2, ... for each detected digit/token.
-        const answers: Record<string, { value: string; confidence: number; blue_pixels: number }> = {};
-        const sourceItems = detectedNumbers.length > 0 ? detectedNumbers : tokens.map((t: any) => t.text);
-        sourceItems.forEach((item: any, i: number) => {
-          const value = typeof item === 'string' ? item : (item.text || '');
-          const conf = typeof item === 'object' && item?.confidence != null ? item.confidence : 0.5;
-          answers[`q_${i + 1}`] = { value: String(value), confidence: Number(conf) || 0.5, blue_pixels: 0 };
-        });
-        // Cleanup temp file
-        try { fs.unlinkSync(tempFilePath); } catch { /* noop */ }
-        return res.json({
-          success: true,
-          isSingleImage: true,
-          answers,
-          ocrAnalysis: {
-            rawOcrText: rawText,
-            extractedTokens: tokens,
-            processingTimeMs: ocrResult?.processingTimeMs || 0,
-            ocrEngine: ocrResult?.evaluation?.ocrEngine || 'EasyOCR (PyTorch Fast Reader)',
-          },
-          totalEvaluated: sourceItems.length,
-        });
-      } catch (e: any) {
-        const raw = e?.message || String(e);
-        // Make the error message useful for the frontend's common-cause hints.
-        // ETIMEDOUT from Node's execFileSync usually means the Python
-        // subprocess was killed at the timeout boundary, not a network blip.
-        const friendly = raw.includes('ETIMEDOUT')
-          ? 'OCR took too long (>60s) and was timed out. The image may be very large or the EasyOCR model is still warming up. Try again.'
-          : raw;
-        return res.status(500).json({ success: false, error: `EasyOCR failed: ${friendly}` });
-      }
-    }
-
-    try {
-      const classes = await dbStore.getClasses();
-      let targetClass = classes.find(c => c.id === classId || c.className.toLowerCase() === String(classId).toLowerCase());
-      if (!targetClass) {
-        const classMatch = String(classId).match(/\d+/);
-        const num = classMatch ? classMatch[0] : '1';
-        targetClass = {
-          id: classId,
-          className: `Class ${num}`,
-          section: 'A',
-          schoolId: '',
-          teacherId: ''
-        };
-      }
-
-      const allStudents = await dbStore.getStudents();
-      let classStudents = allStudents.filter(
-        s => (s.classGroup || '').toLowerCase().includes(targetClass!.className.toLowerCase()) ||
-             targetClass!.className.toLowerCase().includes((s.classGroup || '').toLowerCase())
-      );
-
-      if (classStudents.length === 0) {
-        const classMatch = targetClass.className.match(/\d+/);
-        const classNum = classMatch ? parseInt(classMatch[0], 10) : 1;
-        classStudents = [
-          {
-            id: `STUDENT_PLACEHOLDER_${classNum}`,
-            name: `Student 1 (${targetClass.className})`,
-            age: 7,
-            classGroup: targetClass.className,
-            section: targetClass.section || 'A',
-            schoolId: 'gps-mt-001',
-            currentLevel: classNum * 10,
-            targetLevel: 93,
-            aadharMasked: 'XXXX-XXXX-1234',
-            levelHistory: []
-          }
-        ];
-      }
-
-      // Save PDF or Image file temporarily for Python EasyOCR evaluation
-      const tempDir = path.join(AI_SERVICES_DIR, 'scratch');
-      fs.mkdirSync(tempDir, { recursive: true });
-      const ext = path.extname(filename || 'worksheet.pdf') || '.pdf';
-      const tempFilePath = path.join(tempDir, `scan_${Date.now()}_file${ext}`);
-
-      const cleanBase64 = inputBase64.includes(',') ? inputBase64.split(',')[1] : inputBase64;
-      fs.writeFileSync(tempFilePath, Buffer.from(cleanBase64, 'base64'));
-
-      const classMatch = targetClass.className.match(/\d+/);
-      const classNumber = classMatch ? parseInt(classMatch[0], 10) : 1;
-
-      // Determine which students to evaluate
-      let evalStudents: Student[] = [];
-      if (studentId && studentId !== 'ALL_STUDENTS') {
-        const found = allStudents.find(s => s.id === studentId);
-        if (found) {
-          evalStudents = [found];
-        } else {
-          evalStudents = classStudents.filter(s => s.id === studentId);
-        }
-      } else {
-        evalStudents = classStudents;
-      }
-
-      if (evalStudents.length === 0) {
-        evalStudents = [
-          {
-            id: studentId || `STUDENT_PLACEHOLDER_${classNumber}`,
-            name: `Student (${targetClass.className})`,
-            age: 7,
-            classGroup: targetClass.className,
-            section: targetClass.section || 'A',
-            schoolId: 'gps-mt-001',
-            currentLevel: classNumber * 10,
-            targetLevel: 93,
-            aadharMasked: 'XXXX-XXXX-1234',
-            levelHistory: []
-          }
-        ];
-      }
-
-      // Execute Python EasyOCR ONCE for the uploaded document (Sub-second execution)
-      let sharedOcrResult: any = null;
-      try {
-        const { execFileSync } = await import('child_process');
-        const scriptPath = path.join(AI_SERVICES_DIR, 'scripts', 'ocr.py');
-        const output = execFileSync(PYTHON_BIN, [scriptPath, tempFilePath, studentId || 'ALL_STUDENTS', String(classNumber)], {
-          cwd: AI_SERVICES_DIR,
-          env: { ...process.env, PYTHONIOENCODING: 'utf-8' },
-          timeout: 60000,
-          maxBuffer: 10 * 1024 * 1024
-        });
-        sharedOcrResult = JSON.parse(output.toString());
-      } catch (e: any) {
-        console.warn(`EasyOCR execution info:`, e.message);
-      }
-
-      const results = [];
-      const ocrTokens: Array<{ text: string; confidence: number }> =
-        sharedOcrResult?.evaluation?.extractedTokens ||
-        sharedOcrResult?.extracted_tokens ||
-        sharedOcrResult?.tokens || [];
-
-      const rawOcrText: string =
-        sharedOcrResult?.evaluation?.rawOcrText ||
-        sharedOcrResult?.raw_text ||
-        (ocrTokens.map(t => t.text).join(' ')) || '';
-
-      const realDigits: string[] =
-        sharedOcrResult?.evaluation?.detectedNumbers ||
-        sharedOcrResult?.digits_found ||
-        (rawOcrText.match(/\d+/g) || []);
-
-      for (let sIdx = 0; sIdx < evalStudents.length; sIdx++) {
-        const student = evalStudents[sIdx];
-        const diagQuestions = await dbStore.getStudentAssignedQuestions(student.id, classNumber);
-        const totalQ = (diagQuestions && diagQuestions.length > 0) ? diagQuestions.length : 10;
-        const extractedAnswers: Record<string, string> = {};
-        let score = 0;
-
-        if (diagQuestions.length === 0) {
-          // Fallback mock evaluation if student has no assigned paper yet
-          for (let i = 1; i <= 10; i++) {
-            const qId = `Q_L${(classNumber - 1) * 10 + i}_1`;
-            const digitIndex = (sIdx * 10) + (i - 1);
-            const val = realDigits[digitIndex] !== undefined ? String(realDigits[digitIndex]) : String(i * 2);
-            extractedAnswers[qId] = val;
-            score++;
-          }
-        } else {
-          diagQuestions.forEach((q, idx) => {
-            // Attempt to match extracted digit token for this question index
-            const digitIndex = (sIdx * diagQuestions.length) + idx;
-            const extractedDigit = (realDigits && realDigits[digitIndex] !== undefined)
-              ? String(realDigits[digitIndex]).trim()
-              : null;
-
-            if (extractedDigit !== null) {
-              extractedAnswers[q.question_id] = extractedDigit;
-              if (extractedDigit === String(q.answer).trim()) {
-                score++;
-              }
-            } else {
-              // Fallback match check against raw OCR text
-              const textMatch = rawOcrText.includes(String(q.answer).trim());
-              extractedAnswers[q.question_id] = textMatch ? String(q.answer).trim() : String(q.answer).trim();
-              if (textMatch) score++;
-            }
-          });
-        }
-
-        const percentage = Math.round((score / totalQ) * 100);
-        const recommendedLevel = Math.max(1, Math.min(93, (classNumber - 1) * 10 + Math.ceil(percentage / 10)));
-        const subLevel = percentage >= 80 ? 0 : percentage >= 50 ? 1 : 2;
-
-        const levelHistory = [...(student.levelHistory || []), {
-          level: recommendedLevel,
-          subLevel,
-          date: new Date().toISOString().split('T')[0],
-          reason: CYCLE_NAMES[0] // 'Baseline' - this ICR path is the scanned diagnostic-placement flow
-        }];
-
-        await dbStore.updateStudent(student.id, {
-          currentLevel: recommendedLevel,
-          currentSubLevel: subLevel,
-          targetLevel: Math.min(93, recommendedLevel + 1),
-          levelHistory
-        });
-
-        const report: EvaluationReport = {
-          id: 'rep_icr_file_' + randomUUID().slice(0, 8),
-          studentId: student.id,
-          worksheetId: 'icr_file_scan',
-          score,
-          totalQuestions: diagQuestions.length,
-          conceptMastery: {
-            'Number Sense': percentage >= 70 ? 'Strong' : 'Needs Practice',
-            'Shapes': percentage >= 60 ? 'Strong' : 'Needs Practice',
-            'Operations': percentage >= 50 ? 'Strong' : 'Needs Practice'
-          },
-          narrative: `ICR EasyOCR Answer Sheet Evaluation complete for ${student.name}. Score: ${score}/${diagQuestions.length} (${percentage}%). Assessed at Level ${recommendedLevel}.${subLevel}. Raw OCR: "${rawOcrText.slice(0, 60)}"`,
-          recommendedLevel,
-          recommendedSubLevel: subLevel,
-          timestamp: new Date().toISOString()
-        };
-
-        await dbStore.addEvaluationReport(report);
-
-        results.push({
-          studentId: student.id,
-          studentName: student.name,
-          rollNumber: student.id.slice(-4),
-          score,
-          totalQuestions: diagQuestions.length,
-          percentage,
-          previousLevel: student.currentLevel,
-          newLevel: recommendedLevel,
-          subLevel,
-          questions: diagQuestions.map(q => ({
-            id: q.question_id,
-            question: q.question,
-            correctAnswer: q.answer,
-            topic: q.topic || 'General'
-          })),
-          extractedAnswers,
-          ocrEngine: 'EasyOCR (PyTorch CRAFT Neural Net)',
-          ocrAnalysis: {
-            rawOcrText: rawOcrText || 'Extracted via EasyOCR',
-            extractedTokens: ocrTokens.length > 0 ? ocrTokens : realDigits.map(d => ({ text: d, confidence: 0.95 })),
-            processingTimeMs: sharedOcrResult?.processingTimeMs || 140,
-            ocrEngine: 'EasyOCR (PyTorch CRAFT Neural Net)'
-          },
-          status: percentage >= 50 ? 'Mastered' : 'Needs Remediation'
-        });
-      }
-
-      try { fs.unlinkSync(tempFilePath); } catch { }
-
-      await dbStore.addLog({
-        id: 'log_' + Date.now(),
-        timestamp: new Date().toISOString(),
-        schoolId: targetClass.schoolId,
-        schoolName: targetClass.className,
-        userId: user.id,
-        userEmail: user.email,
-        userRole: user.role,
-        activityType: 'scan',
-        status: 'Success',
-        details: `ICR PDF Scan: Evaluated ${results.length} student answer sheets for ${targetClass.className}`
-      });
-
-      res.json({
-        success: true,
-        isBulk: evalStudents.length > 1,
-        totalEvaluated: results.length,
-        results
-      });
-
-    } catch (err: any) {
-      console.error('ICR PDF Evaluation Error:', err);
-      res.status(500).json({ error: err.message || 'Failed to process ICR PDF scan.' });
-    }
-  });
-
-    // =========================================================================
-  // ICR via external cloud OCR APIs (Google Vision / AWS Textract / Azure /
-  // MiniMax / OCR.space)
-  // =========================================================================
-  // SECURITY: the API key is stored server-side (env var + optional DB
-  // override). The frontend NEVER sees the key — it just picks a provider
-  // and asks the backend to OCR. The user (or admin) configures the key
-  // once via POST /api/icr/cloud-config, and every subsequent scan uses
-  // that server-side key automatically.
-
-  let _cloudKeyCache = null;
-  const getCloudKey = async (provider) => {
-    const envKey = process.env['ICR_CLOUD_API_KEY_' + provider.toUpperCase()];
-    if (envKey) return envKey;
-    if (!_cloudKeyCache) {
-      try {
-        const stored = await dbStore.getConfig('icrCloudKeys');
-        _cloudKeyCache = (stored && typeof stored === 'object') ? stored : {};
-      } catch (_e) {
-        _cloudKeyCache = {};
-      }
-    }
-    return _cloudKeyCache[provider] || null;
-  };
-
-  // Admin endpoint: configure (or clear) a cloud OCR API key.
-  // Restricted to superadmin / admin roles. Returns {provider, configured}.
-  app.post('/api/icr/cloud-config', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-    if (user.role !== 'superadmin' && user.role !== 'admin') {
-      return res.status(403).json({ error: 'Admin role required.' });
-    }
-    const { provider, apiKey } = req.body || {};
-    if (!provider || ['google', 'aws', 'azure', 'minimax', 'ocrspace'].indexOf(provider) === -1) {
-      return res.status(400).json({ error: 'provider must be one of google/aws/azure/minimax/ocrspace.' });
-    }
-    if (!_cloudKeyCache) _cloudKeyCache = {};
-    if (apiKey == null || apiKey === '') {
-      delete _cloudKeyCache[provider];
-    } else {
-      _cloudKeyCache[provider] = apiKey;
-    }
-    try {
-      await dbStore.setConfig('icrCloudKeys', _cloudKeyCache);
-    } catch (e) {
-      return res.status(500).json({ error: 'Failed to persist key: ' + (e && e.message) });
-    }
-    return res.json({
-      success: true,
-      provider: provider,
-      configured: !!_cloudKeyCache[provider],
-    });
-  });
-
-  // Read endpoint: which providers are configured.
-  app.get('/api/icr/cloud-config', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-    const result = {};
-    const providers = ['google', 'aws', 'azure', 'minimax', 'ocrspace'];
-    for (let i = 0; i < providers.length; i++) {
-      const k = await getCloudKey(providers[i]);
-      result[providers[i]] = !!k;
-    }
-    return res.json({ success: true, providers: result });
-  });
-
-  // OCR endpoint: takes {provider, imageDataUrl} only. NO apiKey from frontend.
-  app.post('/api/icr/evaluate-cloud', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { imageDataUrl, provider } = req.body || {};
-    if (!imageDataUrl || typeof imageDataUrl !== 'string' || imageDataUrl.indexOf('data:image/') !== 0) {
-      return res.status(400).json({ error: 'imageDataUrl is required (data URL).' });
-    }
-    if (!provider || ['google', 'aws', 'azure', 'minimax', 'ocrspace'].indexOf(provider) === -1) {
-      return res.status(400).json({ error: 'provider must be one of google/aws/azure/minimax/ocrspace.' });
-    }
-
-    const apiKey = await getCloudKey(provider);
-    if (!apiKey) {
-      return res.status(503).json({
-        error: provider + ' API key not configured on the server. Ask an admin to set it via /api/icr/cloud-config or the ICR_CLOUD_API_KEY_' + provider.toUpperCase() + ' env var.',
-      });
-    }
-
-    // Strip the data URL prefix -> raw base64
-    const commaIdx = imageDataUrl.indexOf(',');
-    const base64Body = commaIdx >= 0 ? imageDataUrl.slice(commaIdx + 1) : imageDataUrl;
-    const t0 = Date.now();
-
-    try {
-      // ===== Google Cloud Vision =====
-      if (provider === 'google') {
-        const visionRes = await fetch(
-          'https://vision.googleapis.com/v1/images:annotate?key=' + encodeURIComponent(apiKey),
-          {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({
-              requests: [{
-                image: { content: base64Body },
-                features: [{ type: 'DOCUMENT_TEXT_DETECTION' }],
-                imageContext: { languageHints: ['en'] },
-              }],
-            }),
-          }
-        );
-        const visionJson = await visionRes.json();
-        if (!visionRes.ok) {
-          const msg = (visionJson && visionJson.error && visionJson.error.message) ||
-            (visionJson && visionJson.responses && visionJson.responses[0] && visionJson.responses[0].error && visionJson.responses[0].error.message) ||
-            ('Google Vision HTTP ' + visionRes.status);
-          return res.status(502).json({ error: 'Google Vision: ' + msg });
-        }
-        const resp = visionJson && visionJson.responses && visionJson.responses[0];
-        if (resp && resp.error) {
-          return res.status(502).json({ error: 'Google Vision: ' + resp.error.message });
-        }
-        const fullText = (resp && resp.fullTextAnnotation && resp.fullTextAnnotation.text) || '';
-        const blocks = (resp && resp.fullTextAnnotation && resp.fullTextAnnotation.pages && resp.fullTextAnnotation.pages[0] && resp.fullTextAnnotation.pages[0].blocks) || [];
-        const tokens = [];
-        for (let bi = 0; bi < blocks.length; bi++) {
-          const paras = blocks[bi].paragraphs || [];
-          for (let pi = 0; pi < paras.length; pi++) {
-            const words = paras[pi].words || [];
-            for (let wi = 0; wi < words.length; wi++) {
-              const word = words[wi];
-              const syms = word.symbols || [];
-              let wtext = '';
-              for (let si = 0; si < syms.length; si++) wtext += (syms[si].text || '');
-              if (!wtext.trim()) continue;
-              const verts = (word.boundingBox && word.boundingBox.vertices) || [];
-              const bbox = [];
-              for (let vi = 0; vi < verts.length; vi++) {
-                bbox.push([verts[vi].x || 0, verts[vi].y || 0]);
-              }
-              if (bbox.length === 0) {
-                bbox.push([0, 0], [0, 0], [0, 0], [0, 0]);
-              }
-              tokens.push({
-                text: wtext,
-                confidence: typeof word.confidence === 'number' ? word.confidence : 0.9,
-                bbox: bbox,
-              });
-            }
-          }
-        }
-        return res.json({
-          success: true,
-          provider: 'google',
-          ocrEngine: 'Google Cloud Vision (DOCUMENT_TEXT_DETECTION)',
-          rawOcrText: fullText,
-          extractedTokens: tokens,
-          processingTimeMs: Date.now() - t0,
-        });
-      }
-
-      // ===== MiniMax (vision-capable chat completion) =====
-      if (provider === 'minimax') {
-        const imageDataUrl = 'data:image/jpeg;base64,' + base64Body;
-        const ocrPrompt =
-          'You are an OCR engine. Read this handwritten answer sheet and ' +
-          'extract every visible mark. For each detected number, symbol, or ' +
-          'word, output one JSON object per line on its own line with the ' +
-          'exact format: {"text": "<exact value>", "confidence": <0..1>}. ' +
-          'Skip printed labels, page numbers, and decorative marks — only ' +
-          'output the handwritten content. Do not include any explanation ' +
-          'or commentary. Output ONLY the JSON lines.';
-        const minimaxRes = await fetch(
-          'https://api.MiniMax.chat/v1/chat/completions',
-          {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-              'Authorization': 'Bearer ' + apiKey,
-            },
-            body: JSON.stringify({
-              model: 'minimax-m3',
-              messages: [{
-                role: 'user',
-                content: [
-                  { type: 'text', text: ocrPrompt },
-                  { type: 'image_url', image_url: { url: imageDataUrl } },
-                ],
-              }],
-              max_tokens: 4096,
-              temperature: 0,
-            }),
-          }
-        );
-        const minimaxJson = await minimaxRes.json();
-        if (!minimaxRes.ok) {
-          const msg = (minimaxJson && minimaxJson.error && minimaxJson.error.message) ||
-            (minimaxJson && minimaxJson.message) ||
-            ('MiniMax HTTP ' + minimaxRes.status);
-          return res.status(502).json({ error: 'MiniMax: ' + msg });
-        }
-        const reply = (minimaxJson && minimaxJson.choices && minimaxJson.choices[0] && minimaxJson.choices[0].message && minimaxJson.choices[0].message.content) || '';
-        const cleaned = String(reply).replace(/\`\`\`json\n?/gi, '').replace(/\`\`\`\n?/g, '').trim();
-        const tokens = [];
-        const lines = cleaned.split('\n');
-        let yPos = 0;
-        for (let li = 0; li < lines.length; li++) {
-          const trimmed = lines[li].trim();
-          if (!trimmed) continue;
-          let parsed = null;
-          try { parsed = JSON.parse(trimmed); } catch (_e) { parsed = null; }
-          if (parsed && parsed.text) {
-            const t = String(parsed.text).trim();
-            const c = typeof parsed.confidence === 'number' ? parsed.confidence : 0.85;
-            if (!t) continue;
-            tokens.push({ text: t, confidence: c, bbox: [[0, yPos], [Math.max(t.length * 12, 30), yPos], [Math.max(t.length * 12, 30), yPos + 24], [0, yPos + 24]] });
-          } else if (trimmed.length > 0 && trimmed.length < 50) {
-            tokens.push({ text: trimmed, confidence: 0.7, bbox: [[0, yPos], [trimmed.length * 12, yPos], [trimmed.length * 12, yPos + 24], [0, yPos + 24]] });
-          }
-          yPos += 30;
-        }
-        const rawText = tokens.map(function (t) { return t.text; }).join(' ');
-        return res.json({
-          success: true,
-          provider: 'minimax',
-          ocrEngine: 'MiniMax minimax-m3 (vision)',
-          rawOcrText: rawText,
-          extractedTokens: tokens,
-          processingTimeMs: Date.now() - t0,
-        });
-      }
-
-      // ===== OCR.space (free tier) =====
-      if (provider === 'ocrspace') {
-        const formBody = new URLSearchParams();
-        formBody.append('base64Image', 'data:image/jpeg;base64,' + base64Body);
-        formBody.append('apikey', apiKey);
-        formBody.append('language', 'eng');
-        formBody.append('isOverlayRequired', 'false');
-        formBody.append('scale', 'true');
-        formBody.append('OCREngine', '2');
-        formBody.append('detectOrientation', 'true');
-        const ocrRes = await fetch('https://api.ocr.space/parse/image', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-          body: formBody.toString(),
-        });
-        const ocrJson = await ocrRes.json();
-        if (ocrJson.IsErroredOnProcessing) {
-          const errMsg = (ocrJson.ErrorMessage && ocrJson.ErrorMessage[0]) ||
-            ocrJson.ErrorDetails ||
-            ('OCR.space HTTP ' + ocrRes.status);
-          return res.status(502).json({ error: 'OCR.space: ' + errMsg });
-        }
-        const parsed = (ocrJson.ParsedResults && ocrJson.ParsedResults[0]) || null;
-        const fullText = (parsed && parsed.ParsedText) || '';
-        // Split on newlines and spaces — synthesize bboxes sequentially top-down.
-        // Use String.prototype.split with a regex — but write the regex with
-        // only \\n to avoid CR/LF ambiguity (OCR.space text uses \\n).
-        const splitRegex = new RegExp(String.fromCharCode(10));
-        const lines = String(fullText).split(splitRegex);
-        const tokens = [];
-        let yPos = 0;
-        for (let li = 0; li < lines.length; li++) {
-          if (!lines[li] || !lines[li].trim()) continue;
-          const words = lines[li].trim().split(/\\s+/);
-          for (let wi = 0; wi < words.length; wi++) {
-            const w = words[wi];
-            if (!w) continue;
-            tokens.push({
-              text: w,
-              confidence: 0.85,
-              bbox: [[0, yPos], [Math.max(w.length * 12, 30), yPos], [Math.max(w.length * 12, 30), yPos + 24], [0, yPos + 24]],
-            });
-          }
-          yPos += 30;
-        }
-        return res.json({
-          success: true,
-          provider: 'ocrspace',
-          ocrEngine: 'OCR.space (Engine 2, free tier)',
-          rawOcrText: fullText,
-          extractedTokens: tokens,
-          processingTimeMs: Date.now() - t0,
-        });
-      }
-
-      // ===== AWS Textract (stub) =====
-      if (provider === 'aws') {
-        return res.status(501).json({
-          error: 'AWS Textract integration is not yet implemented. Pick Google Cloud Vision, MiniMax, OCR.space or use the local OCR button.',
-        });
-      }
-
-      // ===== Azure Computer Vision (stub) =====
-      if (provider === 'azure') {
-        return res.status(501).json({
-          error: 'Azure Computer Vision integration is not yet implemented. Pick Google Cloud Vision, MiniMax, OCR.space or use the local OCR button.',
-        });
-      }
-
-      return res.status(400).json({ error: 'Unknown provider: ' + provider });
-    } catch (e) {
-      return res.status(500).json({ error: 'Cloud OCR failed: ' + (e && e.message ? e.message : String(e)) });
-    }
-  });
+  registerEvaluationRoutes(app);
   registerWorksheetRoutes(app);
-
-  // Submit Completed Worksheet (ICR Structured Ingestion) & Scoring Engine
-  app.post('/api/evaluation/submit', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { worksheetId, studentId, answers } = req.body;
-    if (!worksheetId || !studentId || !answers) {
-      return res.status(400).json({ error: 'Worksheet ID, Student ID, and answer map are required.' });
-    }
-
-    const worksheets = await dbStore.getWorksheets();
-    const ws = worksheets.find(w => w.id === worksheetId);
-    if (!ws) return res.status(404).json({ error: 'Worksheet not found.' });
-
-    const students = await dbStore.getStudents();
-    const student = students.find(s => s.id === studentId);
-    if (!student) return res.status(404).json({ error: 'Student not found.' });
-
-    // Idempotency: a student can only submit a given worksheet once. If this
-    // exact (worksheetId, studentId) pair was already submitted (e.g. the
-    // client retried after a timeout), return the existing result instead of
-    // re-running the AI evaluation, re-mutating the student's level,
-    // and re-appending to level history / delay logs.
-    const existingSubmissions = await dbStore.getAnswerSubmissions();
-    const existingSubmission = existingSubmissions.find(s => s.worksheetId === worksheetId && s.studentId === studentId);
-    if (existingSubmission) {
-      const existingReports = await dbStore.getEvaluationReports();
-      const existingReport = existingReports
-        .filter(r => r.worksheetId === worksheetId && r.studentId === studentId)
-        .sort((a, b) => new Date(b.timestamp).getTime() - new Date(a.timestamp).getTime())[0];
-      return res.json({
-        submission: existingSubmission,
-        report: existingReport,
-        evaluation: existingReport ? {
-          score: existingReport.score,
-          recommendedLevel: existingReport.recommendedLevel,
-          narrative: existingReport.narrative,
-          conceptMastery: existingReport.conceptMastery
-        } : undefined,
-        alreadySubmitted: true
-      });
-    }
-
-    // Handle Timings & Delayed Attempt Escalation (§6.5)
-    const now = new Date();
-    const submissionDeadline = new Date(ws.timing.submissionWindowEnd);
-    const isDelayed = now.getTime() > submissionDeadline.getTime();
-
-    // Grade and generate AI narrative using Gemini AI
-    const studentQuestions = ws.questions.filter(q => q.question_id.startsWith(student.id + '_'));
-    const evaluation = await evaluateAIWorksheet(student.name, student.currentLevel, studentQuestions, answers);
-
-    // Determine subLevel based on question performance at the recommended level
-    let newSubLevel = 0; // default Mastery
-    const recLevel = evaluation.recommendedLevel;
-    const levelQs = studentQuestions.filter(q => q.source_level === recLevel);
-    if (levelQs.length > 0) {
-      let failedCount = 0;
-      levelQs.forEach(q => {
-        const submitted = (answers[q.question_id] || '').trim().toLowerCase();
-        const correct = q.answer.trim().toLowerCase();
-        if (submitted !== correct) failedCount++;
-      });
-      if (failedCount === levelQs.length) {
-        newSubLevel = 2; // Remedial
-      } else if (failedCount > 0) {
-        newSubLevel = 1; // Easier
-      }
-    }
-
-    // Save submission
-    const submission: AnswerSubmission = {
-      id: 'sub_' + student.id + '_' + Date.now(),
-      worksheetId,
-      studentId,
-      studentName: student.name,
-      schoolId: ws.schoolId,
-      classId: ws.classId,
-      submittedAt: now.toISOString(),
-      isDelayed,
-      answers
-    };
-
-    await dbStore.addAnswerSubmission(submission);
-
-    // Save Evaluation Report
-    const report: EvaluationReport = {
-      id: 'rep_' + student.id + '_' + Date.now(),
-      studentId,
-      worksheetId,
-      score: evaluation.score,
-      totalQuestions: studentQuestions.length,
-      conceptMastery: evaluation.conceptMastery,
-      narrative: evaluation.narrative,
-      recommendedLevel: evaluation.recommendedLevel,
-      recommendedSubLevel: newSubLevel,
-      timestamp: now.toISOString()
-    };
-
-    await dbStore.addEvaluationReport(report);
-
-    // If correct, update student levels
-    const levelHistory = [...student.levelHistory];
-    if (evaluation.recommendedLevel !== student.currentLevel || newSubLevel !== (student.currentSubLevel || 0)) {
-      levelHistory.push({
-        level: evaluation.recommendedLevel,
-        subLevel: newSubLevel,
-        date: now.toISOString().split('T')[0],
-        reason: ws.cycle // already one of CYCLE_NAMES
-      });
-    }
-
-    await dbStore.updateStudent(student.id, {
-      currentLevel: evaluation.recommendedLevel,
-      currentSubLevel: newSubLevel,
-      targetLevel: Math.min(93, evaluation.recommendedLevel + 1),
-      levelHistory
-    });
-
-    // Logging & escalation updates if delayed
-    if (isDelayed) {
-      ws.delayLogs.delayedAttemptsCount++;
-      if (!ws.delayLogs.submittingTeachers.includes(user.email)) {
-        ws.delayLogs.submittingTeachers.push(user.email);
-      }
-      await dbStore.updateWorksheet(ws.id, { delayLogs: ws.delayLogs });
-
-      // Increment Teacher's delay count & enforce Defaulter status (§6.5)
-      const users = await dbStore.getUsers();
-      const teacherUser = users.find(u => u.email.toLowerCase() === user.email.toLowerCase());
-      if (teacherUser && teacherUser.role === UserRole.TEACHER) {
-        const curDelays = (teacherUser.delayedAttemptsCount || 0) + 1;
-        const shouldBan = curDelays >= 3;
-        await dbStore.updateUser(teacherUser.id, {
-          delayedAttemptsCount: curDelays,
-          isBanned: shouldBan
-        });
-
-        // Lock school access if all teachers in this school default
-        if (shouldBan && teacherUser.schoolId) {
-          const schoolTeachers = users.filter(u => u.role === UserRole.TEACHER && u.schoolId === teacherUser.schoolId);
-          const allBanned = schoolTeachers.every(t => t.isBanned || t.id === teacherUser.id);
-          if (allBanned && schoolTeachers.length > 0) {
-            await dbStore.updateSchool(teacherUser.schoolId, { isAccessLocked: true });
-          }
-        }
-      }
-
-      await dbStore.addLog({
-        id: 'log_' + Date.now(),
-        timestamp: now.toISOString(),
-        schoolId: ws.schoolId,
-        schoolName: 'GPS',
-        userId: user.id,
-        userEmail: user.email,
-        userRole: user.role,
-        activityType: 'scan',
-        status: 'Delayed',
-        details: `SUBMISSION DELAYED: Answers for ${student.name} uploaded after the 1-hour submission window closed.`
-      });
-    } else {
-      await dbStore.addLog({
-        id: 'log_' + Date.now(),
-        timestamp: now.toISOString(),
-        schoolId: ws.schoolId,
-        schoolName: 'GPS',
-        userId: user.id,
-        userEmail: user.email,
-        userRole: user.role,
-        activityType: 'scan',
-        status: 'Success',
-        details: `Successfully evaluated assessment sheet for ${student.name}.`
-      });
-    }
-
-    // Auto-detect intervention outcomes
-    const interventions = await dbStore.getInterventions();
-    const activeInterventions = interventions.filter(
-      i => i.studentId === studentId && i.status === 'active' && !i.outcome
-    );
-    for (const intv of activeInterventions) {
-      const improved = evaluation.recommendedLevel > intv.currentLevel;
-      await dbStore.updateIntervention(intv.id, {
-        status: 'completed',
-        endDate: now.toISOString().split('T')[0],
-        outcome: {
-          improved,
-          previousLevel: intv.currentLevel,
-          newLevel: evaluation.recommendedLevel,
-          improvementDetails: improved
-            ? `Auto-detected: Student improved from Level ${intv.currentLevel} to Level ${evaluation.recommendedLevel} after intervention targeting ${intv.weakCompetencies.join(', ')}.`
-            : `Auto-detected: Student remained at Level ${intv.currentLevel} after intervention. Further remediation may be needed.`,
-          assessmentId: report.id,
-          detectedAt: now.toISOString()
-        }
-      });
-    }
-
-    res.json({ submission, report, evaluation });
-  });
-
-  // Bulk evaluation reports, scoped identically to GET /api/students (§14).
-  app.get('/api/evaluation/reports', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const [reports, students, schools] = await Promise.all([
-      dbStore.getEvaluationReports(),
-      dbStore.getStudents(),
-      dbStore.getSchools(),
-    ]);
-
-    if (user.role === UserRole.SUPERADMIN) {
-      return res.json(reports);
-    }
-
-    let scopedStudentIds: Set<string>;
-    if (user.role === UserRole.SCHOOL || user.role === UserRole.TEACHER) {
-      scopedStudentIds = new Set(students.filter(s => s.schoolId === user.schoolId).map(s => s.id));
-    } else if (user.role === UserRole.VOLUNTEER) {
-      scopedStudentIds = new Set(students.filter(s => user.assignedSchools?.includes(s.schoolId)).map(s => s.id));
-    } else if (user.role === UserRole.ADMIN || user.role === UserRole.DISTRICT_ADMIN || user.role === UserRole.BLOCK_ADMIN) {
-      const schoolById = new Map(schools.map(sc => [sc.id, sc]));
-      scopedStudentIds = new Set(students.filter(s => {
-        const school = schoolById.get(s.schoolId);
-        if (!school) return false;
-        if (user.role === UserRole.ADMIN) return school.stateCode === user.stateCode;
-        if (user.role === UserRole.DISTRICT_ADMIN) return school.districtCode === user.districtCode;
-        return school.blockCode === user.blockCode; // BLOCK_ADMIN
-      }).map(s => s.id));
-    } else {
-      scopedStudentIds = new Set(students.map(s => s.id));
-    }
-
-    res.json(reports.filter(r => scopedStudentIds.has(r.studentId)));
-  });
-
-  // Evaluation History
-  app.get('/api/evaluation/:studentId/history', async (req, res) => {
-    const reps = await dbStore.getEvaluationReports();
-    const filtered = reps.filter(r => r.studentId === req.params.studentId);
-    res.json(filtered);
-  });
-
-  // Roll up Analytics for Dashboards scoped by Role (§14)
-  app.get('/api/analytics', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const students = await dbStore.getStudents();
-    const schools = await dbStore.getSchools();
-    const worksheets = await dbStore.getWorksheets();
-    const reports = await dbStore.getEvaluationReports();
-
-    // Query params for dynamic filtering
-    const stateCodeParam = (req.query.stateCode as string) || user.stateCode || 'PB';
-    const districtCodeParam = (req.query.districtCode as string) || user.districtCode || 'LDH';
-    const blockCodeParam = (req.query.blockCode as string) || user.blockCode || 'LDH-01';
-
-    // Map geographical attributes to students
-    const studentsWithGeo = students.map(s => {
-      const sch = schools.find(x => x.id === s.schoolId);
-      return {
-        ...s,
-        stateCode: sch?.stateCode || '',
-        districtCode: sch?.districtCode || '',
-        blockCode: sch?.blockCode || ''
-      };
-    });
-
-    const getScopeMetrics = (filteredStudents: typeof studentsWithGeo) => {
-      const count = filteredStudents.length;
-      if (count === 0) {
-        return {
-          avgLevel: 0,
-          certificationRate: 0,
-          topicMastery: {
-            "Number Sense": 0,
-            "Number Operations": 0,
-            "Shapes": 0,
-            "Fractions": 0,
-            "Patterns": 0,
-            "Measurement": 0
-          },
-          levelDistribution: Object.fromEntries(
-            Array.from({ length: 15 }, (_, i) => [`Level ${i + 1}`, 0])
-              .concat([["Level 16+", 0]])
-          )
-        };
-      }
-      const sumLevel = filteredStudents.reduce((acc, s) => acc + s.currentLevel, 0);
-      const avgLevel = Math.round((sumLevel / count) * 10) / 10;
-      const certified = filteredStudents.filter(s => s.currentLevel >= 5).length;
-      const certificationRate = Math.round((certified / count) * 100);
-
-      // Create stable topic mastery values that reflect the current average level
-      const avgCurrentLevel = sumLevel / count;
-      const topicMastery = {
-        "Number Sense": Math.min(100, Math.round(55 + avgCurrentLevel * 8)),
-        "Number Operations": Math.min(100, Math.round(45 + avgCurrentLevel * 9)),
-        "Shapes": Math.min(100, Math.round(58 + avgCurrentLevel * 7)),
-        "Fractions": Math.min(100, Math.round(20 + avgCurrentLevel * 11)),
-        "Patterns": Math.min(100, Math.round(38 + avgCurrentLevel * 10)),
-        "Measurement": Math.min(100, Math.round(32 + avgCurrentLevel * 10))
-      };
-
-      const levelDistribution: Record<string, number> = {};
-      for (let i = 1; i <= 15; i++) {
-        levelDistribution[`Level ${i}`] = filteredStudents.filter((s: any) => s.currentLevel === i).length;
-      }
-      levelDistribution["Level 16+"] = filteredStudents.filter((s: any) => s.currentLevel >= 16).length;
-
-      return {
-        avgLevel,
-        certificationRate,
-        topicMastery,
-        levelDistribution
-      };
-    };
-
-    // Calculate dynamic scopes
-    const national = getScopeMetrics(studentsWithGeo);
-    const state = getScopeMetrics(studentsWithGeo.filter(s => s.stateCode === stateCodeParam));
-    const district = getScopeMetrics(studentsWithGeo.filter(s => s.districtCode === districtCodeParam));
-    const block = getScopeMetrics(studentsWithGeo.filter(s => s.blockCode === blockCodeParam));
-
-    // Compute high-level general metrics
-    const totalStudents = students.length;
-    const totalSchools = schools.length;
-    const totalWorksheets = worksheets.length;
-    const certifiedCount = students.filter(s => s.currentLevel >= 5).length;
-    const certificationPercent = totalStudents > 0 ? Math.round((certifiedCount / totalStudents) * 100) : 0;
-
-    const pipeline = {
-      conducted: worksheets.length * 10,
-      scanned: reports.length,
-      evaluated: reports.length,
-      certified: certifiedCount
-    };
-
-    res.json({
-      totalStudents,
-      totalSchools,
-      totalWorksheets,
-      certificationPercent,
-      pipeline,
-      roleScope: user.role,
-      national,
-      state,
-      district,
-      block
-    });
-  });
-
-  // Comprehensive Super Admin Executive Analytics Endpoint (§ Executive Oversight)
-  // All numbers are computed from live MongoDB Atlas data using FASTER
-  // aggregation helpers on dbStore (countSchoolsFast, countStudentsFast,
-  // countSchoolsByState, etc.) that run a single Mongo $group pipeline
-  // and return only counts — never loads the full 86k+ student / 1.4k+
-  // school documents into memory. Response time: <500ms even on the
-  // full Atlas dataset.
-  app.get('/api/analytics/superadmin', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-    if (user.role !== UserRole.SUPERADMIN && user.role !== UserRole.ADMIN) {
-      return res.status(403).json({ error: 'Forbidden: Superadmin access required.' });
-    }
-
-    try {
-      // Filters
-      const dateRange = (req.query.dateRange as string) || '30d';
-      const stateCode = (req.query.stateCode as string) || 'ALL';
-      const schoolType = (req.query.schoolType as string) || 'ALL';
-      const board = (req.query.board as string) || 'ALL';
-      const grade = (req.query.grade as string) || 'ALL';
-      const status = (req.query.status as string) || 'ALL';
-
-      // Build school filter once, reuse across aggregations
-      const schoolFilter: { stateCode?: string; schoolType?: string; accessLocked?: boolean } = {};
-      if (stateCode !== 'ALL') schoolFilter.stateCode = stateCode;
-      if (schoolType !== 'ALL') schoolFilter.schoolType = schoolType;
-      if (status === 'Active') schoolFilter.accessLocked = false;
-      else if (status === 'Audit Flagged') schoolFilter.accessLocked = true;
-
-      // PARALLEL aggregations — run them concurrently with Promise.all so
-      // the total wall-time is the slowest query, not the sum of all.
-      const [
-        totalSchools,
-        activeSchoolsCount,
-        schoolByState,
-        schoolByType,
-        totalStudents,
-        certifiedCount,
-        studentsBySchool,
-        userCounts,
-        reportStats,
-        totalUsers,
-        allFilteredSchools, // for schoolRankings (we still need names + IDs)
-      ] = await Promise.all([
-        dbStore.countSchoolsFast(schoolFilter),
-        // Active = total when status filter is 'ALL' or 'Active' (FLN doesn't
-        // populate accessLocked on most schools, so treat them as active).
-        // When 'Audit Flagged' is selected, active is 0.
-        dbStore.countSchoolsFast(status === 'Audit Flagged' ? { ...schoolFilter, accessLocked: false } : { ...schoolFilter, accessLocked: { $ne: true } } as any),
-        dbStore.countSchoolsByState(),
-        dbStore.countSchoolsByType(),
-        dbStore.countStudentsFast(),
-        dbStore.countStudentsFast({ currentLevelMin: 5 }),
-        dbStore.getSchoolStudentCounts(),
-        dbStore.countUsersByRole(),
-        dbStore.countReportsByOutcome(),
-        // users count for the KPI tile
-        (async () => {
-          if (dbStore.getDb()) {
-            return await dbStore.getDb()!.collection('users').countDocuments({});
-          }
-          return (dbStore as any).data?.users?.length || 0;
-        })(),
-        // Schools list (still need names + IDs for rankings) — get only
-        // the fields we need, projected to 60-byte records.
-        (async () => {
-          if (dbStore.getDb()) {
-            return await dbStore.getDb()!.collection('schools')
-              .find(buildMongoFilter(schoolFilter))
-              .project({ _id: 1, id: 1, name: 1, stateCode: 1, schoolType: 1 })
-              .toArray() as any[];
-          }
-          let result = (dbStore as any).data?.schools || [];
-          if (stateCode !== 'ALL') result = result.filter((s: any) => s.stateCode === stateCode);
-          if (schoolType !== 'ALL') result = result.filter((s: any) => s.schoolType === schoolType);
-          if (status === 'Active') result = result.filter((s: any) => !s.accessLocked);
-          else if (status === 'Audit Flagged') result = result.filter((s: any) => s.accessLocked);
-          return result.map((s: any) => ({ id: s.id, name: s.name, stateCode: s.stateCode, schoolType: s.schoolType }));
-        })(),
-      ]);
-
-      const auditFlagged = totalSchools - activeSchoolsCount;
-      const certifiedPercent = totalStudents > 0 ? Math.round((certifiedCount / totalStudents) * 100) : 0;
-      const avgScore = reportStats.total > 0 ? reportStats.avgScore : 0;
-      // Map user role counts to dashboard fields
-      const superadmins = userCounts['superadmin'] || 0;
-      const admins = userCounts['admin'] || 0;
-      const districtAdmins = userCounts['district_admin'] || 0;
-      const blockAdmins = userCounts['block_admin'] || 0;
-      const schoolUsers = userCounts['school'] || 0;
-      const teachers = userCounts['teacher'] || 0;
-      const volunteers = userCounts['volunteer'] || 0;
-
-      // State distribution (real counts, not the synthetic 24k/MH style)
-      const stateNamesMap: Record<string, string> = {
-        AN: 'Andaman and Nicobar Islands', AP: 'Andhra Pradesh', AR: 'Arunachal Pradesh', AS: 'Assam',
-        BR: 'Bihar', CH: 'Chandigarh', CG: 'Chhattisgarh', DN: 'Dadra and Nagar Haveli',
-        DD: 'Daman and Diu', DL: 'Delhi NCT', GA: 'Goa', GJ: 'Gujarat', HR: 'Haryana',
-        HP: 'Himachal Pradesh', JK: 'Jammu and Kashmir', JH: 'Jharkhand', KA: 'Karnataka',
-        KL: 'Kerala', LA: 'Ladakh', MP: 'Madhya Pradesh', MH: 'Maharashtra', MN: 'Manipur',
-        ML: 'Meghalaya', MZ: 'Mizoram', NL: 'Nagaland', OD: 'Odisha', PY: 'Puducherry',
-        PB: 'Punjab', RJ: 'Rajasthan', SK: 'Sikkim', TN: 'Tamil Nadu', TS: 'Telangana',
-        TR: 'Tripura', UP: 'Uttar Pradesh', UK: 'Uttarakhand', WB: 'West Bengal',
-      };
-
-      const stateDistribution = schoolByState.map(s => ({
-        stateCode: s.stateCode,
-        stateName: stateNamesMap[s.stateCode] || s.stateCode,
-        schoolsCount: s.count,
-        studentsCount: stateCode === 'ALL'
-          ? (() => {
-              // For ALL, sum studentsBySchool entries whose school is in this state.
-              // We don't have state on studentBySchool key (schoolId only), so
-              // we count from the filtered list: easier to use allFilteredSchools.
-              // For per-state filter, we have the right number already.
-              if (s.stateCode === stateCode) {
-                return totalStudents;
-              }
-              // Approximate: skip the per-state student count when state=ALL
-              // to avoid loading all students. Set to 0 as a placeholder; the
-              // /api/students?stateCode=... endpoint returns accurate counts
-              // when filtered.
-              return 0;
-            })()
-          : (() => {
-              // stateCode is a specific state — count students in schools of
-              // that state. We have allFilteredSchools with stateCode field,
-              // so count students per schoolId in that set.
-              const schIds = new Set(
-                allFilteredSchools.filter((sc: any) => sc.stateCode === s.stateCode)
-                  .map((sc: any) => sc.id)
-              );
-              let count = 0;
-              studentsBySchool.forEach((c, sid) => { if (schIds.has(sid)) count += c; });
-              return count;
-            })(),
-        avgScore: 0,
-      })).sort((a, b) => b.schoolsCount - a.schoolsCount);
-
-      // School type breakdown from real data
-      const performanceBySchoolType = schoolByType.map(t => ({
-        type: t.schoolType || 'Government',
-        avgScore: 0,
-        schoolsCount: t.count,
-      }));
-
-      // Board distribution = school type distribution (FLN doesn't have a
-      // `board` field; schoolType is the closest proxy we can compute live).
-      const boardTotal = schoolByType.reduce((sum, t) => sum + t.count, 0) || 1;
-      const boardDistribution = schoolByType.map(t => ({
-        board: t.schoolType || 'Unknown',
-        schoolsCount: t.count,
-        percentage: Math.round((t.count / boardTotal) * 100),
-      }));
-
-      // Growth trend: 12 months of real new-school cumulative
-      // (we don't track school creation date reliably, so use cumulative
-      // totals bucketed to months — flat per-month for now, but data is
-      // real not synthetic).
-      const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
-      const perMonth = totalSchools / 12;
-      const growthTrend = months.map((m, i) => ({
-        label: m,
-        newSchools: i === 0 ? 0 : Math.round(perMonth / 30),
-        cumulative: Math.min(totalSchools, Math.round(perMonth * (i + 1))),
-      }));
-
-      // School rankings: real filtered schools with computed metrics from
-      // studentsBySchool map (no scan needed).
-      const schoolRankings = allFilteredSchools.map((sch: any) => {
-        const schId = sch.id || sch._id;
-        const totalStud = studentsBySchool.get(schId) || 0;
-        const schStudents = totalStud; // could re-query if we need per-school avg
-        return {
-          rank: 0,
-          id: schId,
-          name: sch.name,
-          stateCode: sch.stateCode,
-          schoolType: sch.schoolType || 'Government',
-          performanceScore: schStudents > 0 ? Math.round((schStudents / 93) * 100) : 0,
-          completionRate: 0,
-          studentSatisfaction: 0,
-          interviewSuccessRate: 0,
-        };
-      });
-      schoolRankings.sort((a, b) => b.performanceScore - a.performanceScore);
-      schoolRankings.forEach((sch, idx) => { sch.rank = idx + 1; });
-
-      // Performance by state
-      const performanceByState = stateDistribution.map(s => ({
-        stateCode: s.stateCode,
-        stateName: s.stateName,
-        avgScore: s.avgScore,
-        prevScore: s.avgScore,
-      }));
-      const topPerformingStates = [...performanceByState].sort((a, b) => b.avgScore - a.avgScore).slice(0, 4);
-      const lowestPerformingStates = [...performanceByState].sort((a, b) => a.avgScore - b.avgScore).slice(0, 4);
-
-      // Interview analytics: real report counts
-      const interviewAnalytics = {
-        totalInterviewsDaily: [],
-        completionRate: reportStats.total > 0 ? 100 : 0,
-        passVsFail: {
-          pass: reportStats.pass,
-          fail: reportStats.fail,
-          passPercent: reportStats.total > 0 ? Math.round((reportStats.pass / reportStats.total) * 100) : 0,
-          failPercent: reportStats.total > 0 ? Math.round((reportStats.fail / reportStats.total) * 100) : 0,
-        },
-        avgDurationMinutes: 0,
-        ratingDistribution: [],
-      };
-
-      // Usage analytics
-      const usageAnalytics = {
-        dailyActiveUsers: 0,
-        weeklyActiveUsers: 0,
-        monthlyActiveUsers: totalUsers,
-        peakLoginHours: [],
-        deviceUsage: { desktop: 0, mobile: 0, tablet: 0 },
-        userByRole: { superadmins, admins, districtAdmins, blockAdmins, schools: schoolUsers, teachers, volunteers },
-      };
-
-      // AI / engagement / system / trends — all 0/empty since FLN doesn't
-      // track these yet, but reported honestly instead of fake numbers.
-      const aiAnalytics = { avgResponseTime: '0s', aiAccuracyScore: 0, avgFeedbackGenTime: '0s', mostAskedDomains: [], mostCommonWeakSkills: [] };
-      const engagementAnalytics = { studentsActiveToday: totalStudents, returningUsersPercentage: 0, newUsersPercentage: 0, dailyEngagementTrend: [] };
-      const systemHealth = { apiUptime: '99.98%', databaseHealth: 'Optimal', activeServers: 'Connected', failedRequests: '0', avgApiLatency: '0ms', errorRate: '0%' };
-      const recentTrends = [
-        { id: 1, type: 'up', title: `Total Students: ${totalStudents.toLocaleString()}`, description: `Across ${totalSchools.toLocaleString()} schools in the system.`, tag: 'Students' },
-        { id: 2, type: 'up', title: `Certified: ${certifiedCount.toLocaleString()} (${certifiedPercent}%)`, description: `Students at FLN level 5 or above.`, tag: 'Outcomes' },
-        { id: 3, type: 'up', title: `Total Users: ${totalUsers.toLocaleString()}`, description: `${superadmins} superadmins, ${admins} admins, ${districtAdmins} district admins, ${blockAdmins} block admins, ${schoolUsers} schools, ${teachers} teachers, ${volunteers} volunteers.`, tag: 'Users' },
-        { id: 4, type: 'star', title: `MongoDB Atlas: Connected`, description: `Live data from ${totalStudents.toLocaleString()} students across ${totalSchools.toLocaleString()} schools.`, tag: 'DB' },
-      ];
-
-      res.json({
-        kpis: {
-          totalRegisteredSchools: totalSchools,
-          activeSchools: activeSchoolsCount,
-          auditFlaggedSchools: auditFlagged,
-          totalStudents,
-          totalCertified: certifiedCount,
-          certifiedPercent,
-          totalTeachers: teachers,
-          totalExamsConducted: reportStats.total,
-          totalInterviewsCompleted: reportStats.total,
-          avgPerformanceScore: avgScore,
-          aiUsageToday: 0,
-        },
-        growthTrend,
-        stateDistribution,
-        boardDistribution,
-        performanceAnalytics: { performanceByState, performanceBySchoolType, topPerformingStates, lowestPerformingStates },
-        interviewAnalytics,
-        usageAnalytics,
-        aiAnalytics,
-        schoolRankings,
-        engagementAnalytics,
-        systemHealth,
-        recentTrends,
-        meta: {
-          appliedFilters: { dateRange, stateCode, schoolType, board, grade, status },
-          generatedAt: new Date().toISOString(),
-          dataSource: 'MongoDB Atlas',
-        },
-      });
-    } catch (err: any) {
-      console.error('[superadmin analytics error]', err);
-      res.status(500).json({ error: 'Failed to compute Super Admin Executive Analytics: ' + (err?.message || 'unknown') });
-    }
-  });
-
-  // Helper: build a MongoDB filter from the schoolFilter object (used to
-  // project the school list to fields we need for the rankings panel).
-  function buildMongoFilter(schoolFilter: { stateCode?: string; schoolType?: string; accessLocked?: boolean }): any {
-    const filter: any = {};
-    if (schoolFilter.stateCode) filter.stateCode = schoolFilter.stateCode;
-    if (schoolFilter.schoolType) filter.schoolType = schoolFilter.schoolType;
-    if (schoolFilter.accessLocked != null) filter.accessLocked = schoolFilter.accessLocked;
-    return filter;
-  }
-
-  // Get active coordinators/administrators
-  // ══════════════════════════════════════════
-  // BULK DIAGNOSTIC GENERATION ENDPOINTS
-  // ══════════════════════════════════════════
-
-  interface BulkDiagnosticJob {
-    jobId: string;
-    classNumber: number;
-    students: Array<{ name: string; studentId: string }>;
-    totalSets: number;
-    completed: number;
-    status: 'running' | 'completed' | 'failed';
-    fileName: string;
-    filePath: string;
-    pdfUrl: string;
-    error: string;
-    startedAt: string;
-    completedAt: string;
-  }
-
-  const bulkJobs = new Map<string, BulkDiagnosticJob>();
-
-  // Start a bulk diagnostic generation job
-  app.post('/api/diagnostic/bulk', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { classNumber, count, students: reqStudents } = req.body;
-
-    // Use provided students array or fall back to count
-    let paperStudents: Array<{ name: string; studentId: string }>;
-    let paperCount: number;
-
-    if (Array.isArray(reqStudents) && reqStudents.length > 0) {
-      paperStudents = reqStudents;
-      paperCount = reqStudents.length;
-    } else {
-      // Automatically fetch real enrolled students for this class from MongoDB
-      const allDbStudents = await dbStore.getStudents();
-      const targetClassName = `Class ${classNumber}`;
-      const enrolled = allDbStudents.filter(s => {
-        const cg = (s.classGroup || '').toLowerCase().trim();
-        return cg === targetClassName.toLowerCase() ||
-               cg === String(classNumber) ||
-               cg.includes(`class ${classNumber}`) ||
-               cg.includes(`class_${classNumber}`);
-      });
-
-      if (enrolled.length === 0) {
-        return res.status(400).json({
-          error: `No enrolled students found in MongoDB for Class ${classNumber}. Please add students to Class ${classNumber} first.`
-        });
-      }
-
-      paperStudents = enrolled.map(s => ({
-        name: s.name,
-        studentId: s.id
-      }));
-      paperCount = paperStudents.length;
-    }
-
-    if (!classNumber) return res.status(400).json({ error: 'classNumber is required.' });
-
-    // Role-based Authorization validation for bulk generation
-    const classes = await dbStore.getClasses();
-    let isAuthorized = false;
-
-    if (user.role === UserRole.SUPERADMIN || user.role === UserRole.ADMIN) {
-      isAuthorized = true;
-    } else if (user.role === UserRole.TEACHER) {
-      isAuthorized = classes.some(c => c.className === `Class ${classNumber}` && (c.teacherId === user.id || c.schoolId === user.schoolId));
-    } else if (user.role === UserRole.VOLUNTEER) {
-      isAuthorized = classes.some(c => c.className === `Class ${classNumber}` && user.assignedSchools?.includes(c.schoolId));
-    } else if (user.role === UserRole.SCHOOL) {
-      isAuthorized = classes.some(c => c.className === `Class ${classNumber}` && c.schoolId === user.schoolId);
-    } else if (user.role === UserRole.BLOCK_ADMIN) {
-      const schools = await dbStore.getSchools();
-      const allowedSchools = schools.filter(s => s.blockCode === user.blockCode).map(s => s.id);
-      isAuthorized = classes.some(c => c.className === `Class ${classNumber}` && allowedSchools.includes(c.schoolId));
-    }
-
-    if (!isAuthorized) {
-      return res.status(403).json({ error: `You are not authorized to generate diagnostic papers for Class ${classNumber}.` });
-    }
-
-    const jobId = 'bulk_' + randomUUID();
-    const job: BulkDiagnosticJob = {
-      jobId,
-      classNumber: Number(classNumber),
-      students: paperStudents,
-      totalSets: paperCount,
-      completed: 0,
-      status: 'running',
-      fileName: '',
-      filePath: '',
-      pdfUrl: '',
-      error: '',
-      startedAt: new Date().toISOString(),
-      completedAt: ''
-    };
-
-    bulkJobs.set(jobId, job);
-
-    // Run in background
-    (async () => {
-      try {
-        const result = await generateDiagnosticPaper({
-          classNumber: job.classNumber,
-          students: paperStudents.map(s => ({ name: s.name, studentId: s.studentId })),
-          onProgress: (setNum, total) => {
-            job.completed = setNum;
-          }
-        });
-
-        job.fileName = result.fileName;
-        job.filePath = result.filePath;
-        job.pdfUrl = `/output/${result.pdfFileName || result.fileName}`;
-        job.status = 'completed';
-        job.completedAt = new Date().toISOString();
-        job.completed = job.totalSets;
-
-        // Store answer keys internally in MongoDB / dbStore mapped strictly per student
-        if (Array.isArray(result.answerKeyData)) {
-          for (const keyItem of result.answerKeyData) {
-            const studentQuestions = (keyItem.questions && keyItem.questions.length > 0)
-              ? keyItem.questions
-              : result.questions;
-
-            await dbStore.addDiagnosticAnswerKey({
-              id: 'dak_' + randomUUID(),
-              jobId: job.jobId,
-              studentId: keyItem.studentId,
-              studentName: keyItem.studentName,
-              classNumber: job.classNumber,
-              setNumber: keyItem.setNum,
-              masterJson: keyItem.masterJson,
-              coords: keyItem.coords,
-              questionPaperJson: keyItem.questionPaperJson,
-              questions: studentQuestions,
-              answerKey: keyItem.answerKey || [],
-              createdAt: new Date().toISOString()
-            });
-
-            if (keyItem.studentId && !keyItem.studentId.startsWith('PLACEHOLDER_')) {
-              await dbStore.assignDiagnosticPaperToStudent(keyItem.studentId, studentQuestions);
-            }
-          }
-        }
-
-        await dbStore.addLog({
-          id: 'log_' + Date.now(),
-          timestamp: new Date().toISOString(),
-          schoolId: user.schoolId || '',
-          schoolName: 'GPS',
-          userId: user.id,
-          userEmail: user.email,
-          userRole: user.role,
-          activityType: 'download',
-          status: 'Success',
-          details: `Bulk diagnostic generated: Class ${classNumber}, ${paperCount} papers`
-        });
-      } catch (err: any) {
-        job.status = 'failed';
-        job.error = err?.message || 'Unknown error during bulk generation';
-        job.completedAt = new Date().toISOString();
-        console.error('Bulk diagnostic job failed:', err);
-      }
-    })();
-
-    res.status(202).json({
-      jobId,
-      classNumber: job.classNumber,
-      totalStudents: paperCount,
-      status: 'running',
-      progressUrl: `/api/diagnostic/bulk/${jobId}/progress`
-    });
-  });
-
-  // Poll bulk job progress
-  app.get('/api/diagnostic/bulk/:jobId/progress', (req, res) => {
-    const job = bulkJobs.get(req.params.jobId);
-    if (!job) return res.status(404).json({ error: 'Job not found.' });
-
-    res.json({
-      jobId: job.jobId,
-      classNumber: job.classNumber,
-      totalStudents: job.totalSets,
-      completed: job.completed,
-      status: job.status,
-      pdfUrl: job.pdfUrl,
-      error: job.error,
-      downloadUrl: job.status === 'completed' ? `/api/diagnostic/bulk/${job.jobId}/download` : null
-    });
-  });
-
-  // Download merged diagnostic PDF
-  app.get('/api/diagnostic/bulk/:jobId/download', (req, res) => {
-    const job = bulkJobs.get(req.params.jobId);
-    if (!job) return res.status(404).json({ error: 'Job not found.' });
-    if (job.status !== 'completed') return res.status(400).json({ error: 'Job not yet completed.' });
-
-    if (!fs.existsSync(job.filePath)) {
-      return res.status(404).json({ error: 'PDF file not found on disk.' });
-    }
-
-    res.download(job.filePath, `class${job.classNumber}_bulk_diagnostic.zip`);
-  });
-
-  // Get stored student diagnostic answer key from MongoDB
-  app.get('/api/diagnostic/student/:studentId/answer-key', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { studentId } = req.params;
-    const { jobId } = req.query;
-
-    try {
-      const answerKey = await dbStore.getStudentDiagnosticAnswerKey(studentId, jobId as string);
-      if (!answerKey) {
-        return res.status(404).json({ error: 'Diagnostic answer key not found for this student.' });
-      }
-      res.json(answerKey);
-    } catch (err: any) {
-      res.status(500).json({ error: err?.message || 'Failed to retrieve answer key.' });
-    }
-  });
-
-  // Generate diagnostic for a single student (enhanced with PDF download)
-  app.post('/api/diagnostic/single', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { studentId, className } = req.body;
-    if (!studentId || !className) {
-      return res.status(400).json({ error: 'studentId and className are required.' });
-    }
-
-    try {
-      const students = await dbStore.getStudents();
-      const student = students.find(s => s.id === studentId);
-      if (!student) return res.status(404).json({ error: 'Student not found.' });
-
-      // Role-based Authorization validation for single generation
-      let isAuthorized = false;
-      if (user.role === UserRole.SUPERADMIN || user.role === UserRole.ADMIN) {
-        isAuthorized = true;
-      } else if (user.role === UserRole.TEACHER) {
-        isAuthorized = student.teacherId === user.id || student.schoolId === user.schoolId;
-      } else if (user.role === UserRole.VOLUNTEER) {
-        isAuthorized = user.assignedSchools?.includes(student.schoolId) || false;
-      } else if (user.role === UserRole.SCHOOL) {
-        isAuthorized = student.schoolId === user.schoolId;
-      } else if (user.role === UserRole.BLOCK_ADMIN) {
-        const schools = await dbStore.getSchools();
-        const school = schools.find(s => s.id === student.schoolId);
-        isAuthorized = school?.blockCode === user.blockCode;
-      }
-
-      if (!isAuthorized) {
-        return res.status(403).json({ error: 'You are not authorized to generate diagnostic papers for this student.' });
-      }
-
-      // Parse class number from classGroup (e.g. "Class 2" -> 2)
-      const classMatch = student.classGroup.match(/\d+/);
-      const classNumber = classMatch ? parseInt(classMatch[0], 10) : 1;
-
-      let questions: Question[];
-      let pdfUrl = '';
-      let useMock = false;
-
-      try {
-        const result = await generateDiagnosticPaper({
-          classNumber,
-          students: [{
-            name: student.name,
-            studentId: student.id,
-            qrData: {
-              age: student.age, classGroup: student.classGroup, section: student.section,
-              schoolId: student.schoolId, currentLevel: student.currentLevel,
-              currentSubLevel: student.currentSubLevel, targetLevel: student.targetLevel
-            }
-          }]
-        });
-        questions = result.questions;
-        pdfUrl = `/output/${result.fileName}`;
-        if (Array.isArray(result.answerKeyData) && result.answerKeyData.length > 0) {
-          const keyItem = result.answerKeyData[0];
-          await dbStore.addDiagnosticAnswerKey({
-            id: 'dak_' + randomUUID(),
-            jobId: 'single_' + student.id,
-            studentId: student.id,
-            studentName: student.name,
-            classNumber,
-            setNumber: 1,
-            masterJson: keyItem.masterJson,
-            coords: keyItem.coords,
-            questionPaperJson: keyItem.questionPaperJson,
-            questions: result.questions,
-            createdAt: new Date().toISOString()
-          });
-        }
-      } catch (err: any) {
-        console.error("Puppeteer paper generation failed, using generateQuestionsForLevel mock:", err);
-        useMock = true;
-        // Generate questions across multiple levels using the level generator
-        const startLevel = (classNumber - 1) * 12 + 1;
-        questions = [];
-        for (let lvl = startLevel; lvl < startLevel + 8; lvl++) {
-          const lvlQuestions = generateQuestionsForLevel(Math.min(lvl, 93), 0);
-          lvlQuestions.forEach(q => {
-            questions.push({
-              ...q,
-              question_id: `DIAG_${lvl}_${q.question_id}`,
-              source_level: Math.min(lvl, 93)
-            });
-          });
-        }
-        // Limit to 12 questions for a reasonable diagnostic
-        questions = questions.slice(0, 12);
-      }
-
-      res.json({
-        student,
-        mockMode: useMock,
-        diagnosticPaper: {
-          id: 'diag_' + student.id + '_' + Date.now(),
-          studentId: student.id,
-          studentName: student.name,
-          questions,
-          pdfUrl
-        }
-      });
-    } catch (err: any) {
-      res.status(500).json({ error: err?.message || 'Failed to generate diagnostic.' });
-    }
-  });
+  registerAnalyticsRoutes(app);
+  registerDiagnosticBulkRoutes(app);
 
   // --- Intervention Tracking & Best Practices Repository ---
